Incremental commit
Adding a lot of random stuff done on the train.  Mostly a bunch of documentation
on the various things in the Django codebase.  I also cleaned up the weird style
I was using in the Javascript to make it a bit more normal.  Some small (and
random) tweaks to naming or making conditionals more verbose.  Need to clean up the
new admin interface (and test in a browser that supports localStorage) and will
then try and migrate to using that.  (the auth system needs to be figured out)